split up web view functionality, pagination is not needed in FXL
